Refactor TradeRepository

+++ MTCG/Repository/TradeRepository.cs
@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Linq;
-using System.Runtime.InteropServices.Marshalling;
-using System.Text;
-using System.Threading.Tasks;
-using MTCG.DAL;
+﻿using MTCG.DAL;
+using MTCG.Interfaces.Logic;
 using MTCG.Interfaces.Repository;
 using MTCG.Logic;
 using MTCG.Models;
 using MTCG.Models.Cards;
 using MTCG.Models.Enums;
+using System.Data;
 
 namespace MTCG.Repository
 {
@@ -34,6 +29,8 @@ namespace MTCG.Repository
         #endregion
 
         private readonly DatabaseService _databaseService = DatabaseService.Instance;
+        private readonly IEventService _eventService = new EventService();
+
 
         public int? AddTradeOfferToDatabase(TradeOffer offer)
         {
@@ -48,10 +45,8 @@ namespace MTCG.Repository
 
                 DatabaseService.AddParameterWithValue(dbCommand, "@userId", DbType.Int32, offer.User.Id);
                 DatabaseService.AddParameterWithValue(dbCommand, "@cardId", DbType.String, offer.Card.Id);
-                DatabaseService.AddParameterWithValue(dbCommand, "@requestedCardType", DbType.String,
-                    offer.RequestedMonster ? "Monster" : "Spell");
+                DatabaseService.AddParameterWithValue(dbCommand, "@requestedCardType", DbType.String, offer.RequestedMonster ? "Monster" : "Spell");
                 DatabaseService.AddParameterWithValue(dbCommand, "@requestedDamage", DbType.Single, offer.RequestedDamage);
-
 
                 // Execute query and error handling
                 int tradeId;
@@ -62,23 +57,24 @@ namespace MTCG.Repository
                 }
                 catch (Exception ex)
                 {
-                    Console.BackgroundColor = ConsoleColor.Red;
-                    Console.ForegroundColor = ConsoleColor.White;
-                    Console.WriteLine(
-                        $"[ERROR] Trade offer of user {offer.User.Username} couldn't be written to the database!");
-                    Console.WriteLine($"[ERROR] {ex.Message}");
-                    Console.ResetColor();
-                    return -1;
+                    _eventService.LogEvent(EventType.Error, $"Couldn't write trade offer of user {offer.User.Username} to database", ex);
+                    return null;
                 }
 
                 return tradeId;
             }
         }
 
-        public bool RemoveTradeDeal(TradeOffer offer)
+
+        public bool RemoveTradeOfferFromDatabase(TradeOffer offer)
         {
             lock (ThreadSync.DatabaseLock)
             {
+                if (offer.Id == null)
+                {
+                    return false;
+                }
+
                 // Prepare SQL query
                 using IDbCommand dbCommand = _databaseService.CreateCommand("""
                                                                       DELETE FROM tradeDeals
@@ -96,17 +92,14 @@ namespace MTCG.Repository
                 }
                 catch (Exception ex)
                 {
-                    Console.BackgroundColor = ConsoleColor.Red;
-                    Console.ForegroundColor = ConsoleColor.White;
-                    Console.WriteLine($"[ERROR] Error deleting trade offer of {offer.User.Username} from the database!");
-                    Console.WriteLine($"[ERROR] {ex.Message}");
-                    Console.ResetColor();
+                    _eventService.LogEvent(EventType.Error, $"Couldn't remove trade offer of User {offer.User.Username} from the database", ex);
                     return false;
                 }
 
                 // Check if at least one row was affected
                 if (rowsAffected <= 0)
                 {
+                    _eventService.LogEvent(EventType.Warning, $"Couldn't remove trade offer of User {offer.User.Username} from the database: Nonexistent Trade Offer", null);
                     return false;
                 }
 
@@ -114,7 +107,8 @@ namespace MTCG.Repository
             }
         }
 
-        public List<TradeOffer> GetAllTradeDeals()
+
+        public List<TradeOffer> GetAllTradeOffers()
         {
             lock (ThreadSync.DatabaseLock)
             {
@@ -124,11 +118,20 @@ namespace MTCG.Repository
                                                                       WHERE active = true
                                                                       """);
 
-                // TODO: Add error handling?
-                using IDataReader reader = dbCommand.ExecuteReader();
+                // Execute query and error handling
+                IDataReader? reader = null;
 
-                List<TradeOffer> deals = new List<TradeOffer>();
+                try
+                {
+                    reader = dbCommand.ExecuteReader();
+                }
+                catch (Exception ex)
+                {
+                    _eventService.LogEvent(EventType.Error, $"Couldn't retrieve list of Trade Offers from database", ex);
+                    return new List<TradeOffer>();
+                }
 
+                List<TradeOffer> offers = new List<TradeOffer>();
 
                 /* Since there is only one database connection that is shared between all threads and queries, two queries cannot run simultaneously
                  * Furthermore, it seems that NPGSQL doesn't support MARS https://learn.microsoft.com/en-us/dotnet/framework/data/adonet/sql/enabling-multiple-active-result-sets
@@ -138,63 +141,54 @@ namespace MTCG.Repository
                 // Check if query returned a result
                 while (reader.Read())
                 {
-                    int tradeId = reader.GetInt32(0);
-                    int userId = reader.GetInt32(1);
-                    string cardId = reader.GetString(2);
-                    string requestedCardType = reader.GetString(3);
-                    float requestedDamage = reader.GetFloat(4);
-
-                    deals.Add(new TradeOffer()
-                        {
-                            Id = tradeId,
-                            User = new User() {Id = userId},
-                            Card = new MonsterCard() {Id = cardId},
-                            RequestedMonster = requestedCardType == "Monster" ? true : false,
-                            RequestedDamage = requestedDamage
-                        });
+                    offers.Add(CreateTradeOfferFromDatabaseEntry(reader));
                 }
 
-                return deals;
+                reader.Close();
+                return offers;
             }
         }
 
-        public TradeOffer? GetTradeDealById(int tradeId)
+
+        public TradeOffer? GetTradeOfferById(int tradeId)
         {
             lock (ThreadSync.DatabaseLock)
             {
                 // Prepare SQL statement
                 using IDbCommand dbCommand = _databaseService.CreateCommand("""
-                                                                      SELECT tradeId, userId, cardId, requestedCardType, requestedElementType, requestedDamage FROM tradeDeals
+                                                                      SELECT tradeId, userId, cardId, requestedCardType, requestedDamage FROM tradeDeals
                                                                       WHERE active = true AND tradeId = @tradeId
                                                                       """);
 
                 DatabaseService.AddParameterWithValue(dbCommand, "@tradeId", DbType.Int32, tradeId);
 
-                // TODO: Add error handling?
-                using IDataReader reader = dbCommand.ExecuteReader();
+                // Execute query and error handling
+                IDataReader? reader = null;
+
+                try
+                {
+                    reader = dbCommand.ExecuteReader();
+                }
+                catch (Exception ex)
+                {
+                    _eventService.LogEvent(EventType.Error, $"Couldn't retrieve trade offer with ID {tradeId} from database", ex);
+                    return null;
+                }
 
                 // Check if query returned a result
                 if (reader.Read())
                 {
-                    int userId = reader.GetInt32(1);
-                    string cardId = reader.GetString(2);
-                    string requestedCardType = reader.GetString(3);
-                    ElementType requestedElementType = Enum.Parse<ElementType>(reader.GetString(4));
-                    float requestedDamage = reader.GetFloat(5);
-
-                    return (new TradeOffer()
-                    {
-                        Id = tradeId,
-                        User = new User() { Id = userId },
-                        Card = new MonsterCard() { Id = cardId },
-                        RequestedMonster = requestedCardType == "Monster" ? true : false,
-                        RequestedDamage = requestedDamage
-                    });
+                    TradeOffer offer = CreateTradeOfferFromDatabaseEntry(reader);
+                    reader.Close();
+                    return offer;
                 }
 
+                _eventService.LogEvent(EventType.Warning, $"Couldn't retrieve trade offer with ID {tradeId} from database: Nonexistent trade offer", null);
+                reader.Close();
                 return null;
             }
         }
+
 
         public TradeOffer? GetTradeDealByCardId(string cardIdToLookup)
         {
@@ -208,31 +202,33 @@ namespace MTCG.Repository
 
                 DatabaseService.AddParameterWithValue(dbCommand, "@cardId", DbType.String, cardIdToLookup);
 
-                // TODO: Add error handling?
-                using IDataReader reader = dbCommand.ExecuteReader();
+                // Execute query and error handling
+                IDataReader? reader = null;
+
+                try
+                {
+                    reader = dbCommand.ExecuteReader();
+                }
+                catch (Exception ex)
+                {
+                    _eventService.LogEvent(EventType.Error, $"Couldn't retrieve trade offer associated with Card {cardIdToLookup} from database", ex);
+                    return null;
+                }
 
                 // Check if query returned a result
                 if (reader.Read())
                 {
-                    int tradeId = reader.GetInt32(0);
-                    int userId = reader.GetInt32(1);
-                    string cardId = reader.GetString(2);
-                    string requestedCardType = reader.GetString(3);
-                    float requestedDamage = reader.GetFloat(4);
-
-                    return (new TradeOffer()
-                    {
-                        Id = tradeId,
-                        User = new User() { Id = userId },
-                        Card = new MonsterCard() { Id = cardId },
-                        RequestedMonster = requestedCardType == "Monster" ? true : false,
-                        RequestedDamage = requestedDamage
-                    });
+                    TradeOffer offer = CreateTradeOfferFromDatabaseEntry(reader);
+                    reader.Close();
+                    return offer;
                 }
 
+                _eventService.LogEvent(EventType.Warning, $"Couldn't retrieve trade offer associated with Card {cardIdToLookup} from database: Nonexistent trade offer", null);
+                reader.Close();
                 return null;
             }
         }
+
 
         public bool SetTradeOfferInactive(int tradeIdToUpdate)
         {
@@ -240,10 +236,10 @@ namespace MTCG.Repository
             {
                 // Prepare SQL Query
                 using IDbCommand dbCommand = _databaseService.CreateCommand("""
-                UPDATE tradeDeals
-                SET active = false
-                WHERE tradeId = @id
-                """);
+                                                                            UPDATE tradeDeals
+                                                                            SET active = false
+                                                                            WHERE tradeId = @id
+                                                                            """);
 
                 DatabaseService.AddParameterWithValue(dbCommand, "@id", DbType.Int32, tradeIdToUpdate);
 
@@ -254,16 +250,36 @@ namespace MTCG.Repository
                 }
                 catch (Exception ex)
                 {
-                    Console.BackgroundColor = ConsoleColor.Red;
-                    Console.ForegroundColor = ConsoleColor.White;
-                    Console.WriteLine($"[ERROR] Trade offer with ID {tradeIdToUpdate} couldn't be set to inactive");
-                    Console.WriteLine($"[ERROR] {ex.Message}");
-                    Console.ResetColor();
+                    _eventService.LogEvent(EventType.Error, $"Couldn't set trade offer with ID {tradeIdToUpdate} to inactive", ex);
                     return false;
                 }
 
                 return true;
             }
+        }
+
+
+        /// <summary>
+        /// creates a TradeOffer from a single database entry in the `tradeDeals` table
+        /// </summary>
+        /// <param name="reader">IDataReader of a query which is currently reading an entry from the `tradeDeals` table</param>
+        /// <returns>TradeOffer object containing the information stored in the database</returns>
+        private TradeOffer CreateTradeOfferFromDatabaseEntry(IDataReader reader)
+        {
+            int tradeId = reader.GetInt32(0);
+            int userId = reader.GetInt32(1);
+            string cardId = reader.GetString(2);
+            string requestedCardType = reader.GetString(3);
+            float requestedDamage = reader.GetFloat(4);
+
+            return (new TradeOffer()
+            {
+                Id = tradeId,
+                User = new User() { Id = userId },
+                Card = new MonsterCard() { Id = cardId },
+                RequestedMonster = requestedCardType == "Monster" ? true : false,
+                RequestedDamage = requestedDamage
+            });
         }
     }
 }
